requirements/local.txt
# =============================================================================
# SD LMS - Local Development Requirements
# =============================================================================

-r base.txt

# Debug
django-debug-toolbar>=4.3.0
django-silk>=5.1.0
ipdb>=0.13.13
ipython>=8.23.0

# Testing
pytest>=8.1.1
pytest-django>=4.8.0
pytest-cov>=5.0.0
pytest-xdist>=3.5.0
pytest-asyncio>=0.23.6
factory-boy>=3.3.0
faker>=24.9.0
freezegun>=1.4.0
responses>=0.25.0
pytest-mock>=3.14.0

# Code Quality
ruff>=0.4.1
mypy>=1.9.0
django-stubs>=4.2.7
djangorestframework-stubs>=3.14.5
pre-commit>=3.7.0

# Type checking
types-redis>=4.6.0.20240417
types-pytz>=2024.1.0.20240417
types-Pillow>=10.2.0.20240415

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.5.18
mkdocstrings[python]>=0.24.3

# Profiling
py-spy>=0.3.14
memory-profiler>=0.61.0
line-profiler>=4.1.2

# Shell
django-extensions>=3.2.3
Werkzeug>=3.0.2
watchdog>=4.0.0

# API Testing
httpx>=0.27.0
